Changed embedded linkedin link
The embedded linked in link was stretching out my contancts page on
mobile so I switched it to a simpler link.

Bootstrap also seems to not be working correctly locally but it is
unchanged since the last push yet it works fine on the server. This is
to test if this new push will also break bootstrap server side.