Comments can now be submitted on the Index page
A create comment view was added which allows the user to input a
username and comment. This will post the comment model created from the
view to the homecontroller, through the message broker api and finally
into the message broker service which adds it to the database. The page
then refreshes and the comment is displayed.