Each view given a controller. Can now comment on each page.
Visitors to the page can now comment on each page. This was achieved
through usings my settings interface in the comment controller. Settings
for lastView and lastController were added. The settings were injected
into the controller. When a user clicks on the submit a comment link on
a certain page, two strings are passed from the view to the controller
based on the page the link was clicked on. So i pass the name of the
page and the name of the controller for that page. This allows me to set
them in my settings and later use them when a comment is submitted to
set the webpage value of the comment before it gets sent off to the
message broker. Now comments will appear on the page the link was
clicked on only.  I then use the lastView and lastController settigns to
redirect the user to the page they came from so they can see their
comment submitted.

I am also aware that if a user just enters the url for the create
comment page without clicking the link on any of the pages, their
comment will not submit properly, it will not show up and the page will
not redirect.