Comments Rendered Partial
The CreateComment view and view of the list of comments are now embedded
into each page.

I removed most of the hack job I did in the last commit to get comments
to work on each page and display only that pages comments.

I added a new method to the messagebroker api which grabs only the
comments for the page that requests them. So all the comments are no
longer sent to the view and sorted in the view.

The comment model now has a constructor so the webpage I am posting a
comment on can easily be passed and set before the comment is sent to
the message broker.