Removed formatting
Removed the formatting from the comments list as it is causing more
problems than it is making the comments look nice. This is also not
something I need to be spending time on at the moment so it has been
removed for now. The comment usernames and comments were overlapping in
mobile view.