Changed how the test works to avoid inappropriate failures/passes
A random number is now added to the Webpage string. This is used to find
the newly submitted test comment without having to reset the message
broker service to avoid finding a previously submitted test comment.
Resulting in a pass if the current test did not submit a comment
correctly.

checking for the any comments in the 3rd step before asigning the var
actualComment ensures that an exception is not thrown but an assertion
fails and the test ends.